agri-advisor/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = ""
authors = [
    { name = "rwik04", email = "rwik04@kgpian.iitkgp.ac.in" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.29,<0.4.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "requests (>=2.32.4,<3.0.0)",
    "tavily-python (>=0.7.10,<0.8.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "langgraph (>=0.6.4,<0.7.0)",
    "aiohttp (>=3.9.0,<4.0.0)",
    "qdrant-client[fastembed] (>=1.15.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "gradio-client (>=1.11.1,<2.0.0)",
    "gradio (>=5.42.0,<6.0.0)",
    "openai (>=1.99.9,<2.0.0)",
    "trafilatura (>=2.0.0,<3.0.0)"
]

[tool.poetry]
packages = [{ include = "agents", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
